% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_FLAV_INTAKE.R
\name{DII_FLAV_INTAKE}
\alias{DII_FLAV_INTAKE}
\title{DII_FLAV_INTAKE: Calculate the Flavonoids Component within the DII}
\usage{
DII_FLAV_INTAKE(FLAV_DR1TOT_PATH = NULL, FLAV_DR2TOT_PATH = NULL)
}
\arguments{
\item{FLAV_DR1TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 1. The file names should be formatted like: flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc.}

\item{FLAV_DR2TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 2. The file names should be formatted like: flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc.}
}
\value{
A dataframe with the variable name SEQN and variable names for the intakes of isoflavones, anthocyanidins, flavan-3-ols, flavanones, flavones and flavonols for Day 1, Day2, or both.
}
\description{
To calculate the intakes of flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for participants from NHANES 2007-2010 and 2017-2018.
}
\examples{
# Example 1
DII_FLAV_INTAKE(FLAV_DR1TOT_PATH = './flav_dr1tot_0710.sas7bdat',
                FLAV_DR2TOT_PATH = './flav_dr2tot_0710.sas7bdat')
# Example 2
DII_FLAV_INTAKE(FLAV_DR1TOT_PATH = './flav_dr1tot_0710.sas7bdat')
# Example 3
DII_FLAV_INTAKE(FLAV_DR2TOT_PATH = flav_dr2tot_1718) # flav_dr2tot_1718 is a dataframe already imported into the Global Environment from file flav_dr2tot_1718.sas7bdat.
}
