% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_GARLIC_INTAKE.R
\name{DII_GARLIC_INTAKE}
\alias{DII_GARLIC_INTAKE}
\title{DII_GARLIC_INTAKE: Calculate the Garlic Component within the DII}
\usage{
DII_GARLIC_INTAKE(
  FLATTENED_FILEPATH = NULL,
  DR1IFF_PATH = NULL,
  DR2IFF_PATH = NULL
)
}
\arguments{
\item{FLATTENED_FILEPATH}{A dataframe generated by the 'FNDDSSRLinks_FNDDSIngred_flattened' function to flatten a FNDDSSRLinks/FNDDSIngred file within a certain version of FNDDS.}

\item{DR1IFF_PATH}{The file path for the Individual Foods File for Day 1. The file names should be formatted like: DR1IFF_E.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR1IFF_F.XPT, DRXIFF_B.XPT, etc.}

\item{DR2IFF_PATH}{The file path for the Individual Foods File for Day 2. The file names should be formatted like: DR2IFF_C.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR2IFF_D.XPT, DRXIFF_B.XPT, etc.}
}
\value{
A dataframe with the variable name SEQN, and one or both of DR1TGARLIC and DR2TGARLIC. For DRXIFF_B, the variable name is DRXTGARLIC.
}
\description{
To calculate the garlic component within the Dietary Inflammatory Index (DII) for NHANES participants (NHANES 2001-2020), it is not recommended to use NHANES 1999-2000 due to the absence of an FNDDS file for that survey period. Recipes used to calculate garlic intake are detailed in the file titled 'Recipes used to calculate the intakes of garlic, ginger, onion, pepper, thyme, and oregano.xlsx', which can be downloaded from the following link: \url{https://github.com/zxucmu/dietaryindexNDP/tree/Supplements/Example\%20files\%20for\%20version\%201.0.0}.
}
\examples{
# Example 1
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_4.1_flattened,
                  DR1IFF_PATH = './DR1IFF_E.XPT',
                  DR2IFF_PATH = './DR2IFF_E.XPT') # fnddsSRlinks_4.1_flattened is a dataframe generated by: fnddsSRlinks_4.1_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS4.mdb').
# Example 2
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1314_flattened,
                  DR1IFF_PATH = dr1iff.h) # fnddsSRlinks_1314_flattened is a dataframe generated by: fnddsSRlinks_1314_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS2013-2014.mdb'); dr1iff.h is a dataframe already imported into the Global Environment from file DR1IFF_H.XPT.
# Example 3
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsIngred_1718_flattened,
                  DR2IFF_PATH = './DR2IFF_J.XPT') # fnddsIngred_1718_flattened is a dataframe generated by: fnddsIngred_1718_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS_2017-2018.mdb').
# Example 4
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR1IFF_PATH = './DRXIFF_B.XPT',
                  DR2IFF_PATH = drxiff.b) # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb'); drxiff.b is a dataframe already imported into the Global Environment from file DRXIFF_B.XPT.
# Example 5
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR1IFF_PATH = './DRXIFF_B.XPT') # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb').
# Example 6
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR2IFF_PATH = './DRXIFF_B.XPT') # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb').
# Example 7
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR1IFF_PATH = drxiff,
                  DR2IFF_PATH = drxiff) # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb'); drxiff is a dataframe already imported into the Global Environment from file DRXIFF.XPT.
# Example 8
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR1IFF_PATH = './DRXIFF.XPT') # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb').
# Example 9
DII_GARLIC_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                  DR2IFF_PATH = './DRXIFF.XPT') # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb').
}
