% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_NHANES_CALCULATION.R
\name{DII_NHANES_CALCULATION}
\alias{DII_NHANES_CALCULATION}
\title{DII_NHANES_CALCULATION: Incorporating Additional Components to Calculate the DII}
\usage{
DII_NHANES_CALCULATION(
  FPED_PATH = NULL,
  NUTRIENT_PATH = NULL,
  DEMO_PATH = NULL,
  FPED_PATH2 = NULL,
  NUTRIENT_PATH2 = NULL,
  OTHER_INGREDIENTS = NULL
)
}
\arguments{
\item{FPED_PATH}{The file path for the total Food Patterns Equivalents Intakes per Individual for Day 1. The file name should be formatted like: fped_dr1tot_0708.sas7bdat, pyr_tot.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from such files.}

\item{NUTRIENT_PATH}{The file path for the total Nutrient Intakes file for Day 1. The file name should be formatted like: DR1TOT_J.XPT, DRXTOT_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR1TOT_J.XPT, DRXTOT_B.XPT, etc.}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be formatted like: DEMO_J.XPT. Alternatively, it can be a dataframe already imported into the Global Environment from files like DEMO_J.XPT.}

\item{FPED_PATH2}{The file path for the total Food Patterns Equivalents Intakes per Individual for Day 2. The file name should be formatted like: fped_dr2tot_1718.sas7bdat, pyr_tot_d2.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like fped_dr2tot_1718.sas7bdat, pyr_tot_d2.sas7bdat, etc.}

\item{NUTRIENT_PATH2}{The file path for the total Nutrient Intakes file for Day 2. The file name should be formatted like: DR2TOT_D.XPT, DRXTOT_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR2TOT_D.XPT, DRXTOT_B.XPT, etc.}

\item{OTHER_INGREDIENTS}{A dataframe generated by the DII_ONESTEP_INTAKE() function that includes SEQN and the intakes of garlic, ginger, onion, pepper, oregano, thyme, flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for Day 1, Day 2, or both.}
}
\value{
The DII and its component scores: DII is the total score (the sum of all components), while DII_NOETOH is the score without alcohol.
}
\description{
This function was developed based on the DII_NHANES_FPED() from the R package 'dietaryindex'. The calculation of the dietary inflammation index (DII) incorporates garlic, ginger, onion, pepper, and oregano/thyme in addition to the components used in the DII_NHANES_FPED() function for NHANES participants. Additionally, the intakes of flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones are included for participants from NHANES 2007-2010 and 2017-2018.
}
\examples{
# Example 1
DII_NHANES_CALCULATION(FPED_PATH = fped_dr1tot_0910.sas7bdat,
                       NUTRIENT_PATH = DR1TOT_F.XPT,
                       DEMO_PATH = DEMO_F.XPT,
                       FPED_PATH2 = fped_dr2tot_0910.sas7bdat,
                       NUTRIENT_PATH2 = DR2TOT_F.XPT,
                       OTHER_INGREDIENTS = dii.nhanes.calculation.f) # dii.nhanes.calculation.f is a dataframe generated by the 'DII_ONE_INTAKE' function that includes SEQN and the intakes of garlic, ginger, onion, pepper, oregano, thyme, flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for Day 1 and Day 2.
# Example 2
DII_NHANES_CALCULATION(FPED_PATH = fped_dr1tot_0708.sas7bdat,
                       NUTRIENT_PATH = dr1tot.e,
                       DEMO_PATH = DEMO_E.XPT,
                       OTHER_INGREDIENTS = dii.nhanes.calculation.e) # dr1tot.e is a dataframe already imported into the Global Environment from file DR1TOT_E.XPT; dii.nhanes.calculation.e is a dataframe generated by the 'DII_ONE_INTAKE' function that includes SEQN and the intakes of garlic, ginger, onion, pepper, oregano, thyme, flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for Day 1 and Day 2.
# Example 3
DII_NHANES_CALCULATION(DEMO_PATH = DEMO_J.XPT,
                       FPED_PATH2 = fped_dr2tot_1718,
                       NUTRIENT_PATH2 = DR2TOT_J.XPT,
                       OTHER_INGREDIENTS = dii.nhanes.calculation.j) # fped_dr2tot_1718 is a dataframe already imported into the Global Environment from file fped_dr2tot_0708.sas7bdat; dii.nhanes.calculation.j is a dataframe generated by the 'DII_ONE_INTAKE' function that includes SEQN and the intakes of garlic, ginger, onion, pepper, oregano, thyme, flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for Day 1 and Day 2.
# Example 4
DII_NHANES_CALCULATION(FPED_PATH = fped_dr1tot_0304.sas7bdat,
                       NUTRIENT_PATH = DR1TOT_C.XPT,
                       DEMO_PATH = demo.c) # demo.c is a dataframe already imported into the Global Environment from file DEMO_C.XPT.
# Example 5
DII_NHANES_CALCULATION(FPED_PATH = pyr_tot.sas7bdat,
                       NUTRIENT_PATH = DRXTOT_B.XPT,
                       DEMO_PATH = DEMO_B.XPT)
}
