% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_NHANES_PLUS.R
\name{DII_NHANES_PLUS}
\alias{DII_NHANES_PLUS}
\title{DII_NHANES_PLUS: Calculate the DII with Additional Components Included in One Step}
\usage{
DII_NHANES_PLUS(
  FPED_PATH = NULL,
  NUTRIENT_PATH = NULL,
  DEMO_PATH = NULL,
  FPED_PATH2 = NULL,
  NUTRIENT_PATH2 = NULL,
  FNDDS_PATH = NULL,
  DR1IFF_PATH = NULL,
  DR2IFF_PATH = NULL,
  FLAV_DR1TOT_PATH = NULL,
  FLAV_DR2TOT_PATH = NULL
)
}
\arguments{
\item{FPED_PATH}{The file path for the total Food Patterns Equivalents Intakes per Individual for Day 1. The file name should be formatted like: fped_dr1tot_0708.sas7bdat, pyr_tot.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like fped_dr1tot_0708.sas7bdat, pyr_tot.sas7bdat, etc.}

\item{NUTRIENT_PATH}{The file path for the total Nutrient Intakes file for Day 1. The file name should be formatted like: DR1TOT_J.XPT, DRXTOT_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR1TOT_J.XPT, DRXTOT_B.XPT, etc.}

\item{DEMO_PATH}{The file path for the DEMOGRAPHIC data. The file name should be formatted like: DEMO_J.XPT. Alternatively, it can be a dataframe already imported into the Global Environment from files like DEMO_J.XPT.}

\item{FPED_PATH2}{The file path for the total Food Patterns Equivalents Intakes per Individual for Day 2. The file name should be formatted like: fped_dr2tot_1718.sas7bdat, pyr_tot_d2.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like fped_dr2tot_1718.sas7bdat, pyr_tot_d2.sas7bdat, etc.}

\item{NUTRIENT_PATH2}{The file path for the total Nutrient Intakes file for Day 2. The file name should be formatted like: DR2TOT_D.XPT, DRXTOT_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR2TOT_D.XPT, DRXTOT_B.XPT, etc.}

\item{FNDDS_PATH}{The file path for the a FNDDS file in Microsoft Access® format. The file names should be formatted like: FNDDS.mdb, FNDDS_2017-2018.mdb, etc. Alternatively, it can be a dataframe that has already been imported into the Global Environment from the table named FNDDSSRLinks or FNDDSIngred within an FNDDS file in Microsoft Access® format, such as FNDDS.mdb, FNDDS_2017-2018.mdb, etc.}

\item{DR1IFF_PATH}{The file path for the Individual Foods File for Day 1. The file names should be formatted like: DR1IFF_E.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR1IFF_F.XPT, DRXIFF_B.XPT, etc.}

\item{DR2IFF_PATH}{The file path for the Individual Foods File for Day 2. The file names should be formatted like: DR2IFF_C.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR2IFF_D.XPT, DRXIFF_B.XPT, etc.}

\item{FLAV_DR1TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 1. The file names should be formatted like: flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc.}

\item{FLAV_DR2TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 2. The file names should be formatted like: flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc.}
}
\value{
A dataframe with the variable name SEQN and variable names for the intakes of garlic, ginger, onion, pepper, thyme/oregano, isoflavones, anthocyanidins, flavan-3-ols, flavanones, flavones and flavonols for Day 1, Day 2, or both.
}
\description{
This function calculates the dietary inflammation index (DII) within one step, incorporating garlic, ginger, onion, pepper, and oregano/thyme, along with the components used in the DII_NHANES_FPED() from the R package 'dietaryindex' for NHANES participants. Additionally, for participants from NHANES 2007-2010 and 2017-2018, the intakes of flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones are included.
}
\examples{
#' # Example 1
DII_NHANES_PLUS(FPED_PATH = './fped_dr1tot_0910.sas7bdat',
                NUTRIENT_PATH = './DR1TOT_F.XPT',
                DEMO_PATH = './DEMO_F.XPT',
                FPED_PATH2 = './fped_dr2tot_0910.sas7bdat',
                NUTRIENT_PATH2 = './DR2TOT_F.XPT', FNDDS_PATH = './FNDDS5.mdb',
                DR1IFF_PATH = './DR1IFF_E.XPT', DR2IFF_PATH = './DR2IFF_E.XPT',
                FLAV_DR1TOT_PATH = './flav_dr1tot_0710.sas7bdat',
                FLAV_DR2TOT_PATH = './flav_dr2tot_0710.sas7bdat')
# Example 2
DII_NHANES_PLUS(FPED_PATH = fped_dr1tot_0708,
                NUTRIENT_PATH = './DR1TOT_E.XPT',
                DEMO_PATH = './DEMO_E.XPT',
                FNDDS_PATH = './FNDDS4.1.mdb',
                DR1IFF_PATH = './DR1IFF_E.XPT',
                FLAV_DR1TOT_PATH = './flav_dr1tot_0708.sas7bdat') # fped_dr1tot_0708 is a dataframe already imported into the Global Environment from file fped_dr1tot_0708.sas7bdat.
# Example 3
DII_NHANES_PLUS(DEMO_PATH = demo.j,
                FPED_PATH2 = './fped_dr2tot_1718.sas7bdat',
                NUTRIENT_PATH2 = './DR2TOT_J.XPT',
                FNDDS_PATH = './FNDDS_2017-2018.mdb',
                DR2IFF_PATH = './DR2IFF_J.XPT',
                FLAV_DR2TOT_PATH = './flav_dr2tot_1718.sas7bdat') # demo.j is a dataframe already imported into the Global Environment from file DEMO_J.XPT.
# Example 4
DII_NHANES_PLUS(FPED_PATH = './pyr_tot_d1.sas7bdat',
                NUTRIENT_PATH = dr1tot.c,
                DEMO_PATH = './DEMO_C.XPT',
                FNDDS_PATH = './FNDDS2.mdb',
                DR1IFF_PATH = './DR1IFF_C.XPT') # dr1tot.c is a dataframe already imported into the Global Environment from file DR1TOT_C.XPT.
# Example 5
DII_NHANES_PLUS(FPED_PATH2 = './fped_dr2tot_1314.sas7bdat',
                NUTRIENT_PATH2 = './DR2TOT_H.XPT',
                DEMO_PATH = './DEMO_H.XPT',
                FNDDS_PATH = './FNDDS2013-2014.mdb',
                DR2IFF_PATH = dr2iff.h) # dr2iff.h is a dataframe already imported into the Global Environment from file DR2IFF_H.XPT.
# Example 6
DII_NHANES_PLUS(FPED_PATH = './pyr_tot.sas7bdat',
                NUTRIENT_PATH = './DRXTOT_B.XPT',
                DEMO_PATH = './DEMO_B.XPT',
                FNDDS_PATH = './FNDDS.mdb',
                DR1IFF_PATH = './DRXIFF_B.XPT')
}
