% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_ONESTEP_INTAKE.R
\name{DII_ONESTEP_INTAKE}
\alias{DII_ONESTEP_INTAKE}
\title{DII_ONESTEP_INTAKE: Calculate Additional Components within the DII in One Step}
\usage{
DII_ONESTEP_INTAKE(
  FLATTENED_FILEPATH = NULL,
  DR1IFF_PATH = NULL,
  DR2IFF_PATH = NULL,
  FLAV_DR1TOT_PATH = NULL,
  FLAV_DR2TOT_PATH = NULL
)
}
\arguments{
\item{FLATTENED_FILEPATH}{A dataframe generated by the 'FNDDSSRLinks_FNDDSIngred_flattened' function to flatten a FNDDSSRLinks/FNDDSIngred file within a certain version of FNDDS.}

\item{DR1IFF_PATH}{The file path for the Individual Foods File for Day 1. The file names should be formatted like: DR1IFF_E.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR1IFF_F.XPT, DRXIFF_B.XPT, etc.}

\item{DR2IFF_PATH}{The file path for the Individual Foods File for Day 2. The file names should be formatted like: DR2IFF_C.XPT, DRXIFF_B.XPT, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like DR2IFF_D.XPT, DRXIFF_B.XPT, etc.}

\item{FLAV_DR1TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 1. The file names should be formatted like: flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr1tot_0710.sas7bdat, flav_dr1tot_1718.sas7bdat, etc.}

\item{FLAV_DR2TOT_PATH}{The file path for the Flavonoid Intake Data File for Day 2. The file names should be formatted like: flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc. Alternatively, it can be a dataframe already imported into the Global Environment from files like flav_dr2tot_0708.sas7bdat, flav_dr2tot_1718.sas7bdat, etc.}
}
\value{
A dataframe with the variable name SEQN and variable names for the intakes of garlic, ginger, onion, pepper, thyme/oregano, isoflavones, anthocyanidins, flavan-3-ols, flavanones, flavones and flavonols for Day 1, Day 2, or both.
}
\description{
To calculate the garlic, ginger, onion, pepper, and thyme/oregano components within the Dietary Inflammatory Index(DII) for NHANES participants (NHANES 2001-2020), as well as flavan-3-ols, flavones, flavonols, flavonones, anthocyanidins, and isoflavones for NHANES 2007-2010 and 2017-2018. It is not recommended to calculate the intakes of garlic, ginger, onion, pepper, and thyme/oregano for NHANES 1999-2000 due to the absence of an FNDDS file for that survey period. Recipes used to calculate these intakes are detailed in the file titled 'Recipes used to calculate the intakes of garlic, ginger, onion, pepper, thyme, and oregano.xlsx', which can be downloaded from the following link: \url{https://github.com/zxucmu/dietaryindexNDP/blob/04cf7e1799353e3de465882cb7176b4007bf34d8/Recipes\%20used\%20to\%20calculate\%20the\%20intakes\%20of\%20garlic\%20ginger\%20onion\%20pepper\%20thyme\%20and\%20oregano.xlsx}.
}
\examples{
# Example 1
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_4.1_flattened,
                   DR1IFF_PATH = './DR1IFF_E.XPT',
                   DR2IFF_PATH = './DR2IFF_E.XPT',
                   FLAV_DR1TOT_PATH = './flav_dr1tot_0708.sas7bdat',
                   FLAV_DR2TOT_PATH = './flav_dr2tot_0708.sas7bdat') # fnddsSRlinks_4.1_flattened is a dataframe generated by: fnddsSRlinks_4.1_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS4.mdb').
# Example 2
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1314_flattened,
                   DR1IFF_PATH = './DR1IFF_H.XPT',
                   DR1IFF_PATH = './DR2IFF_H.XPT') # fnddsSRlinks_1314_flattened is a dataframe generated by: fnddsSRlinks_1314_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS2013-2014.mdb').
# Example 3
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1314_flattened,
                   DR1IFF_PATH = dr1iff.h) # fnddsSRlinks_1314_flattened is a dataframe generated by: fnddsSRlinks_1314_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS2013-2014.mdb'); dr1iff.h is a dataframe already imported into the Global Environment from file DR1IFF_H.XPT.
# Example 4:
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsIngred_1516_flattened,
                   DR2IFF_PATH = './DR2IFF_I.XPT') # fnddsIngred_1516_flattened is a dataframe generated by: fnddsIngred_1516_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS_2015-2016.mdb').
# Example 5
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_5.0_flattened,
                   DR1IFF_PATH = './DR1IFF_F.XPT',
                   FLAV_DR1TOT_PATH = flav_dr1tot_0710) # fnddsSRlinks_5.0_flattened is a dataframe generated by: fnddsSRlinks_5.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS5.mdb'); flav_dr1tot_0710 is a dataframe already imported into the Global Environment from file flav_dr1tot_0708.sas7bdat.
# Example 6
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_5.0_flattened,
                   DR2IFF_PATH = dr2iff.f,
                   FLAV_DR2TOT_PATH = './flav_dr2tot_0710.sas7bdat') # fnddsSRlinks_5.0_flattened is a dataframe generated by: fnddsSRlinks_5.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS5.mdb'); dr2iff.f is a dataframe already imported into the Global Environment from file DR2IFF_F.XPT.
# Example 7
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                   DR1IFF_PATH = './DRXIFF_B.XPT',
                   DR2IFF_PATH = drxiff.b) # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb'); drxiff.b is a dataframe already imported into the Global Environment from file DRXIFF_B.XPT.
# Example 8
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                   DR1IFF_PATH = drxiff.b) # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb'); drxiff.b is a dataframe already imported into the Global Environment from file DRXIFF_B.XPT.
# Example 9
DII_ONESTEP_INTAKE(FLATTENED_FILEPATH = fnddsSRlinks_1.0_flattened,
                   DR2IFF_PATH = './DRXIFF_B.XPT') # fnddsSRlinks_1.0_flattened is a dataframe generated by: fnddsSRlinks_1.0_flattened <- FNDDSSRLinks_FNDDSIngred_flattened(FNDDS = './FNDDS.mdb').
}
