% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DII_ONESTEP_INTAKE_RESULT.R
\name{DII_ONESTEP_INTAKE_RESULT}
\alias{DII_ONESTEP_INTAKE_RESULT}
\title{DII_ONESTEP_INTAKE_RESULT: Retrieve Pre-computed Additional Components within the DII Calculation}
\usage{
DII_ONESTEP_INTAKE_RESULT(
  NHANESCycle = c("0102", "0304", "0506", "0708", "0910", "1112", "1314", "1516", "1718",
    "1720"),
  DAY = c("first", "second", "both")
)
}
\arguments{
\item{NHANESCycle}{A character string indicating the NHANES cycle (e.g., "0102", "0304").}

\item{DAY}{A character string indicating which day's result to retrieve ('first', 'second', 'both').}
}
\value{
A data frame containing the additional components within the DII calculation for the specified NHANES cycle. If the specified cycle is not found, an error message will be returned.
}
\description{
This function retrieves pre-computed results from the DII_ONESTEP_INTAKE() function, allowing researchers to extract additional components within the Dietary Inflammatory Index(DII) calculation for specific NHANES cycles as needed for their analysis.
}
\examples{
DII_ONESTEP_INTAKE_RESULT(NHANESCycle = "1112")
}
