% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNDDSSRLinks_FNDDSIngred_flattened.R
\name{FNDDSSRLinks_FNDDSIngred_flattened}
\alias{FNDDSSRLinks_FNDDSIngred_flattened}
\title{To flatten a FNDDSSRLinks or FNDDSIngred file within a specific version of the USDA Food and Nutrient Database for Dietary Studies (FNDDS)}
\usage{
FNDDSSRLinks_FNDDSIngred_flattened(FNDDS_PATH = NULL)
}
\arguments{
\item{FNDDS_PATH}{The file path for a FNDDS file in Microsoft Access® format. The file names should be formatted like: FNDDS.mdb, FNDDS_2017-2018.mdb, etc. Alternatively, it can be a dataframe that has already been imported into the Global Environment from the table named FNDDSSRLinks or FNDDSIngred within an FNDDS file in Microsoft Access® format, such as FNDDS.mdb, FNDDS_2017-2018.mdb, etc. To obtain FNDDS files, please refer to the following link: http://www.ars.usda.gov/ba/bhnrc/fsrg.}
}
\value{
A dataframe generated by this function to flatten a FNDDSSRLinks/FNDDSIngred file within a certain version of FNDDS.
}
\description{
This function is to flatten a FNDDSSRLinks or FNDDSIngred file within a specific version of the FNDDS to ensure that all FNDDS food codes are ultimately linked to SR/Ingredient codes. The proportional weight contribution of all SR/Ingredient codes linked to each individual FNDDS food is also calculated.
}
\examples{
# Example 1
FNDDSSRLinks_FNDDSIngred_flattened(FNDDS_PATH = './FNDDS.mdb')
# Example 2
FNDDSSRLinks_FNDDSIngred_flattened(FNDDS_PATH = FNDDSSRLinks_2.0) # FNDDSSRLinks_2.0 is a dataframe that imported into the Global Environment from the table named FNDDSSRLinks within the FNDDS2.mdb file.
# Example 3
FNDDSSRLinks_FNDDSIngred_flattened(FNDDS_PATH = './FNDDS2013-2014.mdb')
# Example 4
FNDDSSRLinks_FNDDSIngred_flattened(FNDDS_PATH = FNDDSIngred_1718) # FNDDSIngred_1718 is a dataframe that has been imported into the Global Environment from the table named FNDDSIngred within FNDDS_2017-2018.mdb.
}
